+++ CHANGELOG.md
@@ -4,9 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 ## [Unreleased]
+### Added
+- The reload command has returned. It allows for runtime game config reloading
+
 
 ## [0.3.2] - 2019-06-30
-## Fixed
+### Fixed
 - Regexps for games were incorrectly named in the config